# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LeadForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/sessions:
    get:
      operationId: getChatSessions
      tags: [chat]
      summary: Get chat sessions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create a chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /chat/sessions/{id}/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get messages for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message and get AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessageResponse"
...
components:
  schemas:
...
    ChatSession:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        status:
          type: string
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, title, status, messageCount, createdAt, updatedAt]
    ChatSessionInput:
      type: object
      properties:
        title:
          type: string
      required: [title]
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
        content:
          type: string
        cards:
          type: ["string", "null"]
        actionType:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, sessionId, role, content, createdAt]
    ChatMessageInput:
      type: object
      properties:
        content:
          type: string
        cardResponse:
          type: string
      required: [content]
    ChatMessageResponse:
      type: object
      properties:
        userMessage:
          $ref: "#/components/schemas/ChatMessage"
        assistantMessage:
          $ref: "#/components/schemas/ChatMessage"
        autoProceeds:
          type: boolean
        autoProceedSeconds:
          type: integer
      required: [userMessage, assistantMessage, autoProceeds]
...