# lib/api-zod/src/generated/types/chatSession.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LeadForge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatSession {
  id: number;
  userId: number;
  title: string;
  status: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}